Change Eclipse settings; Create definitions for new methods and refresh old ones
Rewritten all code regarding drawing a Scene. Created new getters. Changed a little approach with Actors - now there are no ActorsCopies. All  actors have iterators pointing to a Model and an ShaderProgram which are all stored separately in a Scene. So now Actor is just a collection of settings and "instructions" about how to render another instance of a Model (with a given shader)